Update /run and add tests
